fuck you too

+++ Content.Goobstation.Shared/Wraith/Minions/Harbinger/SpikerShuffleSystem.cs
@@ -33,7 +33,8 @@ public sealed class SpikerShuffleSystem : EntitySystem
             _statusOld.TryRemoveStatusEffect(ent.Owner, statusEffect);
 
         // Shitcode edit here, knockdownStatusEffect And StunnedStatusEffect
-        EntityManager.RemoveComponents(ent.Owner, ent.Comp.ComponentsToRemove);
+        if (ent.Comp.ComponentsToRemove != null)
+            EntityManager.RemoveComponents(ent.Owner, ent.Comp.ComponentsToRemove);
 
         _statusNew.TryAddStatusEffect(ent.Owner, ent.Comp.StatusEffect, out _, ent.Comp.Duration);
         _statusNew.TryAddStatusEffect(ent.Owner, ent.Comp.StatusAbilityDisable, out _, ent.Comp.Duration); // disable using actions
